[DROOLS-3603] Showing a warning popup on collection removal.
