+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/main/java/org/drools/workbench/screens/scenariosimulation/client/popup/ConfirmPopupView.java
@@ -21,7 +21,7 @@ import org.jboss.errai.ui.shared.api.annotations.Templated;
 
 @Dependent
 @Templated
-public class ConfirmPopupView extends ScenarioConfirmationPopupView implements ConfirmPopup {
+public class ConfirmPopupView extends AbstractScenarioConfirmationPopupView implements ConfirmPopup {
 
     @Override
     public void show(String mainTitleText, String mainText) {
