+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/main/java/org/drools/workbench/screens/scenariosimulation/client/popup/ScenarioConfirmationPopupView.html
@@ -17,6 +17,7 @@
                     <h2 data-field="main-question"></h2>
                     <div>
                         <p data-field="text-1"><p/>
+                        <p data-field="text-question"></p>
                     </div>
                 </div>
                 <div class="modal-footer">
