+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/main/java/org/drools/workbench/screens/scenariosimulation/client/popup/YesNoConfirmPopupPresenter.java
@@ -27,7 +27,7 @@ import org.uberfire.mvp.Command;
 public class YesNoConfirmPopupPresenter implements YesNoConfirmPopup.Presenter {
 
     @Inject
-    YesNoConfirmPopupView yesNoConfirmPopupView;
+    protected YesNoConfirmPopupView yesNoConfirmPopupView;
 
     @Override
     public void show(final String title,
