+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/test/java/org/drools/workbench/screens/scenariosimulation/client/popup/ScenarioConfirmationPopupViewPresenterTest.java
@@ -16,8 +16,10 @@
 
 package org.drools.workbench.screens.scenariosimulation.client.popup;
 
+import com.ait.lienzo.test.LienzoMockitoTestRunner;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.mockito.Mock;
 
 import static org.mockito.Matchers.eq;
@@ -25,8 +27,9 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-//@RunWith(LienzoMockitoTestRunner.class)
-public abstract class ScenarioConfirmationPopupViewPresenterTest extends AbstractDeletePopupViewTest {
+
+@RunWith(LienzoMockitoTestRunner.class)
+public class ScenarioConfirmationPopupViewPresenterTest extends AbstractDeletePopupViewTest {
 
     @Mock
     private ScenarioConfirmationPopupView scenarioConfirmationPopupViewMock;
@@ -37,7 +40,7 @@ public abstract class ScenarioConfirmationPopupViewPresenterTest extends Abstrac
     public void setup() {
         scenarioConfirmationPopupPresenter = spy(new ScenarioConfirmationPopupPresenter() {
             {
-                this.scenarioConfirmationPopupView = ScenarioConfirmationPopupViewPresenterTest.this.scenarioConfirmationPopupViewMock;
+                this.scenarioConfirmationPopupView = scenarioConfirmationPopupViewMock;
             }
         });
     }
