+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/test/java/org/drools/workbench/screens/scenariosimulation/client/popup/ScenarioConfirmationPopupViewTest.java
@@ -16,110 +16,30 @@
 
 package org.drools.workbench.screens.scenariosimulation.client.popup;
 
-import com.google.gwt.dom.client.ParagraphElement;
-import org.jboss.errai.common.client.dom.HTMLElement;
-import org.jboss.errai.common.client.dom.MouseEvent;
-import org.jboss.errai.ui.client.local.spi.TranslationService;
+import com.ait.lienzo.test.LienzoMockitoTestRunner;
 import org.junit.Before;
-import org.junit.Test;
-import org.mockito.Mock;
-import org.uberfire.client.views.pfly.widgets.Button;
+import org.junit.runner.RunWith;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyString;
-import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.reset;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.spy;
 
-public abstract class ScenarioConfirmationPopupViewTest extends AbstractDeletePopupViewTest {
-
-    protected ScenarioConfirmationPopupView popupView;
-
-    @Mock
-    protected ParagraphElement text1Mock;
-
-    @Mock
-    protected ParagraphElement textQuestionMock;
-
-    @Mock
-    protected Button cancelButtonMock;
-
-    @Mock
-    protected Button okDeleteButtonMock;
-
-    @Mock
-    protected MouseEvent mouseEventMock;
-
-    @Mock
-    protected TranslationService translationServiceMock;
+@RunWith(LienzoMockitoTestRunner.class)
+public class ScenarioConfirmationPopupViewTest extends AbstractScenarioConfirmationPopupViewTest {
 
     @Before
     public void setup() {
         super.commonSetup();
+        popupView = spy(new ScenarioConfirmationPopupView() {
+            {
+                this.mainTitle = mainTitleMock;
+                this.mainQuestion = mainQuestionMock;
+                this.text1 = text1Mock;
+                this.textQuestion = textQuestionMock;
+                this.cancelButton = cancelButtonMock;
+                this.okDeleteButton = okDeleteButtonMock;
+                this.modal = modalMock;
+                this.translationService = translationServiceMock;
+            }
+        });
     }
 
-    @Test
-    public void init() {
-        popupView.init();
-        verify(cancelButtonMock, times(1)).setText(anyString());
-    }
-
-    @Test
-    public void show() {
-        popupView.show(MAIN_TITLE_TEXT,
-                       MAIN_QUESTION_TEXT,
-                       TEXT1_TEXT,
-                       TEXT_QUESTION_TEXT,
-                       OKDELETE_BUTTON_TEXT,
-                       okDeleteCommandMock);
-        verifyShow(MAIN_TITLE_TEXT,
-                   MAIN_QUESTION_TEXT,
-                   TEXT1_TEXT,
-                   TEXT_QUESTION_TEXT);
-        assertEquals(okDeleteCommandMock, popupView.okDeleteCommand);
-    }
-
-    @Test
-    public void getElement() {
-        final HTMLElement retrieved = popupView.getElement();
-        assertNotNull(retrieved);
-    }
-
-    @Test
-    public void hide() {
-        popupView.hide();
-        verify(modalMock, times(1)).hide();
-    }
-
-    @Test
-    public void onOkDeleteClick() {
-        popupView.okDeleteCommand = null;
-        popupView.onOkDeleteClick(mouseEventMock);
-        verify(okDeleteCommandMock, never()).execute();
-        verify(popupView, times(1)).hide();
-        reset(popupView);
-        popupView.okDeleteCommand = okDeleteCommandMock;
-        popupView.onOkDeleteClick(mouseEventMock);
-        verify(okDeleteCommandMock, times(1)).execute();
-        verify(popupView, times(1)).hide();
-    }
-
-    @Test
-    public void onCancelClick() {
-        popupView.onCancelClick(mouseEventMock);
-        verify(popupView, times(1)).hide();
-    }
-
-    protected void verifyShow(String mainTitleText, String mainQuestionText, String text1Text, String textQuestionText) {
-        verify(popupView, times(1)).conditionalShow(eq(mainTitleMock), eq(mainTitleText));
-        verify(popupView, times(1)).conditionalShow(eq(mainQuestionMock), eq(mainQuestionText));
-        verify(popupView, times(1)).conditionalShow(eq(text1Mock), eq(text1Text));
-        verify(popupView, times(1)).conditionalShow(eq(textQuestionMock), eq(textQuestionText));
-        verify(popupView, times(1)).conditionalShow(eq(okDeleteButtonMock), any(), any());
-        verify(modalMock, times(1)).show();
-    }
 }
